Setup logging on tests
